Fix unittest-based tests with assume context
The pytest hook to handle failed assumptions is
not called for unittest-based tests. This leads
to an issue where failed assumptions are ignored,
and if you use the context manager version, assertions
are swallowed & not re-raised.

Fix is to use pytest_runtest_call hook instead, which
is used for both unittest & regular pytest tests.